requirements.txt
streamlit>=1.24.0
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
openai>=1.3.7
scikit-learn>=1.3.0
httpx>=0.24.1
plotly>=5.15.0
pythainlp>=5.0.0
urllib3>=2.0.7
tenacity>=8.0.0
